requirements.txt
libsql_experimental==0.0.40
python-dotenv==1.0.1
Requests==2.32.3
